Add QLSTM algorithm to the project
A new file called QLSTM.py has been added under the algorithms folder. This file includes implementation of QLSTM algorithm which is an approach to reinforcement learning that uses LSTM for handling sequential data. Initiating this algorithm also required a minor modification in the __init__.py file to import the newly added algorithm.